+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -  Validation report is now saved as a PDF file
 -  Add button to run all automatable checks at once
 -  Add button to uncheck all previously checked checklist items
+-  Add json schema file with the checklist definition - This allows validation of user-created checklists
+-  Add icon to processing provider and algorithms
 
 
 ## [0.4.0] - 2020-06-25